# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Keep this title stable because it controls generated import paths.
  title: Api
  version: 1.0.0
  description: QuizIt competitive aptitude practice API
servers:
  - url: /api
    description: QuizIt API
...
paths:
...
  /topics/{topic_id}/subtopics:
    get:
      operationId: listSubtopics
      tags: [catalog]
      summary: List subtopics in a topic
      parameters:
        - $ref: "#/components/parameters/TopicId"
      responses:
        "200":
          description: Subtopics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subtopic"
  /questions:
    get:
      operationId: listQuestions
      tags: [catalog]
      summary: Browse active questions
      parameters:
        - name: topic_id
          in: query
          schema:
            type: integer
        - name: subtopic_id
          in: query
          schema:
            type: integer
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Questions without answer keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
  /questions/{question_id}:
    get:
      operationId: getQuestion
      tags: [catalog]
      summary: Get one active question
      parameters:
        - $ref: "#/components/parameters/QuestionId"
      responses:
        "200":
          description: Question without answer key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    TopicId:
      name: topic_id
      in: path
      required: true
      schema:
        type: integer
    QuestionId:
      name: question_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Subtopic:
      type: object
      required: [id, topic_id, name, slug]
      properties:
        id: { type: integer }
        topic_id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: [string, "null"] }
    Question:
      type: object
      required: [id, subtopic_id, text, options]
      properties:
        id: { type: string }
        subtopic_id: { type: integer }
        text: { type: string }
        text_html: { type: [string, "null"] }
        directions_html: { type: [string, "null"] }
        options: {}
        options_html: {}
        answer_letter: { type: [string, "null"] }
        explanation: { type: [string, "null"] }
        explanation_html: { type: [string, "null"] }
        difficulty: { type: [string, "null"] }
...