# lib/api-zod/src/generated/types/question.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * QuizIt competitive aptitude practice API
 * OpenAPI spec version: 1.0.0
 */

export interface Question {
  id: string;
  subtopic_id: number;
  text: string;
  /** @nullable */
  text_html?: string | null;
  /** @nullable */
  directions_html?: string | null;
  options: unknown;
  options_html?: unknown;
  /** @nullable */
  answer_letter?: string | null;
  /** @nullable */
  explanation?: string | null;
  /** @nullable */
  explanation_html?: string | null;
  /** @nullable */
  difficulty?: string | null;
}